+++ ossh_loot.go
