+++ utils/payload_hash.go
@@ -1,4 +1,4 @@
-package main
+package utils
 
 import (
 	"strings"
